Fixes non-determinism in functions calling the function pykpn.common.mapping.Mapping.from_list. This function is now deprecated and a warning is given for it. The old functionality is moved now to a from_list_random function in the same class, which makes its non-determinism explicit in the name.
Implements a deterministic full mapper ComFullMapper with best effort communication selection. This should
be used instead of the old from_list for deteriministic outcomes. Also makes multiple functions from the
deterministic mappers involved static.

The representations now use this new full mapper. Fixed signature of representations allowing not
to pass a config (as it was an optional keyword arg).