Removed file comparison for hog-big-3.csv test.
The problem with the test that there are different valid optimal
outputs, since sthe application have similar deadlines. The problem
might also occurs with other tests and scheduler, so it might make
sense either to make more flexible validatio, or to output just the
final energy consumption of the schedules, which should not vary.