make sure that simulation objects are pickable and usable with multiprocessing
This requires stripping them of the simpy environment. In order to achieve
this, this patch makes BaseSimulation usable within env block. When entering
the block, the envirionment and other required objects are created, and when
exiting the block, they are deleted. This ensures that the environment is not
part of the simulation object outside of the with block and ensures that it can
be pickled.