platform: odroid: added extra static power of the platform.
The current power model includes only the power dissipated by the processors.
However, there are components consuming the power. To make the results
closer to the real platform, I added "extra_static_power" field to the
odroid configuration.
The model power does not yet include the power of the fan.